requirements.txt
python-telegram-bot==20.8
Flask==2.2.5
aiohttp==3.8.4
pip install python==3.11.9